ref(sampling): Remove activate modal
